% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETo_BR.R
\name{ETo_BR}
\alias{ETo_BR}
\title{Função para calcular ETo FAO 56 das estações automaticas do Brasil}
\usage{
ETo_BR(date)
}
\arguments{
\item{date}{uma data}
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
Está função serve para calcular a ETo FAO 56 no Brasil
}
\author{
Santos Henrique Brant Dias
}
