% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc.sec.dy.rad.extr.R
\name{inc.sec.dy.rad.extr}
\alias{inc.sec.dy.rad.extr}
\title{Função para acrescentar raster radiação extraterrestre do dia}
\usage{
inc.sec.dy.rad.extr(datfram, stackday, shapflbr)
}
\arguments{
\item{datfram}{um Data Frame gerado com a função ETo_BR do pacote com os as coordenadas x e y escritas Lat e Long especificamente, além da data!}

\item{stackday}{um stack raster}

\item{shapflbr}{um arquivo shapfile load rgdal}
}
\value{
Returns a stack raster day modeling
}
\description{
Esta função serve criar raster de radiação extraterrestre
}
\author{
Santos Henrique Brant Dias
}
