% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indcs.R
\name{indcs}
\alias{indcs}
\title{Função para calculo dos indices de avaliação do desepenho de modelos}
\usage{
indcs(prd, obs)
}
\arguments{
\item{prd}{: numeric 'data.frame', 'matrix' or 'vector' with simulated values}

\item{obs}{: numeric 'data.frame', 'matrix' or 'vector' with observed values}
}
\value{
Returns a data.frame with the AWS data requested
}
\description{
Esta função serve para calcular os indices de desempenho de modelos
}
\author{
Santos Henrique Brant Dias
}
