% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_raster_EToBR.R
\name{pred.raster.EToBR}
\alias{pred.raster.EToBR}
\title{Função gera o raster da modelagem de ETo para todo o Brasil}
\usage{
pred.raster.EToBR(date, stacktotal, stackmonth, stackday, shapflbr, model)
}
\arguments{
\item{date}{um data para modelagem}

\item{stacktotal}{um stack raster}

\item{stackmonth}{um stack raster}

\item{stackday}{um stack raster}

\item{shapflbr}{um arquivo shapfile load rgdal}

\item{model}{um modelo para realização da modelagem e criaçõ do mapa raster, caracter baseado no pacote caret}
}
\value{
Returns a stack raster day modeling
}
\description{
Esta função serve para gerar o raster de ETo para o Brasil
}
\author{
Santos Henrique Brant Dias
}
